* commit 'fc5a93f7a2ea2be740f2c909e8a6a192fa65adc8':
  Fix incorrect parameter types for locale funcs.

+++ libc/bionic/strtoll_l.cpp
@@ -28,6 +28,6 @@
 
 #include <stdlib.h>
 
-long long strtoll_l(const char *nptr, char **endptr, size_t base, locale_t) {
+long long strtoll_l(const char *nptr, char **endptr, int base, locale_t) {
   return strtoll(nptr, endptr, base);
 }
